Extract AppLockedScreen.kt to another file + minor refactor (& fix bug caused by the previous refactor)
